Refactor scrapers with better imports
- Also bugfix composite scraper run
- Add "scraper" and "scraper_config" to the metadata in composite
  scraper
