+++ akd/tools/scrapers/__init__.py
@@ -1,18 +1,15 @@
-from .pdf_scrapers import PDFScraperToolInputSchema, SimplePDFScraper
-from .web_scrapers import (
-    Crawl4AIWebScraper,
-    SimpleWebScraper,
-    WebpageScraperToolConfig,
-    WebpageScraperToolInputSchema,
-    WebpageScraperToolOutputSchema,
-)
+from ._base import ScraperToolInputSchema, ScraperToolOutputSchema
+from .omni import DoclingScraper, DoclingScraperConfig
+from .pdf_scrapers import PDFScraperInputSchema, SimplePDFScraper
+from .web_scrapers import Crawl4AIWebScraper, SimpleWebScraper
 
 __all__ = [
     "SimplePDFScraper",
     "SimpleWebScraper",
     "Crawl4AIWebScraper",
-    "PDFScraperToolInputSchema",
-    "WebpageScraperToolInputSchema",
-    "WebpageScraperToolOutputSchema",
-    "WebpageScraperToolConfig",
+    "PDFScraperInputSchema",
+    "ScraperToolInputSchema",
+    "ScraperToolOutputSchema",
+    "DoclingScraper",
+    "DoclingScraperConfig",
 ]
