+++ akd/tools/scrapers/web_scrapers.py
@@ -1,300 +1,16 @@
-import json
 import re
-from typing import Any, Dict, List, Literal, Optional, Tuple
 
-import requests
+import httpx
 from bs4 import BeautifulSoup
 from crawl4ai import AsyncWebCrawler
-from docling.datamodel.base_models import InputFormat
-from docling.datamodel.pipeline_options import PdfPipelineOptions, TableFormerMode
-from docling.document_converter import (
-    DocumentConverter,
-    HTMLFormatOption,
-    PdfFormatOption,
-)
-from docling_core.types import DoclingDocument
-from loguru import logger
 from markdownify import markdownify
-from pydantic import Field, HttpUrl, field_validator, model_validator
 from readability import Document
 from requests import HTTPError, RequestException
 
-from akd._base import InputSchema, OutputSchema
-from akd.structures import SearchResultItem
-from akd.tools import BaseTool, BaseToolConfig
-
-from .utils import _DoclingMetadataExtractor
+from ._base import ScraperToolInputSchema, ScraperToolOutputSchema, WebScraper
 
 
-class WebpageScraperToolInputSchema(InputSchema):
-    """
-    Input schema for the WebpageScraperTool.
-    """
-
-    url: HttpUrl = Field(
-        ...,
-        description="URL of the webpage to scrape.",
-    )
-    include_links: bool = Field(
-        default=True,
-        description="Whether to preserve hyperlinks in the markdown output.",
-    )
-
-
-class WebpageMetadata(SearchResultItem):
-    keywords: Optional[List[str]] = Field(
-        None,
-        description="List of keywords from the webpage content.",
-    )
-
-
-class WebpageScraperToolOutputSchema(OutputSchema):
-    """Schema for the output of the WebpageScraperTool."""
-
-    content: str = Field(
-        ...,
-        description="The scraped content in markdown format.",
-    )
-    metadata: WebpageMetadata = Field(
-        ...,
-        description="Metadata about the scraped webpage.",
-    )
-
-
-class WebpageScraperToolConfig(BaseToolConfig):
-    """Configuration for the WebpageScraperTool."""
-
-    user_agent: str = Field(
-        default=(
-            "Mozilla/5.0 (Windows NT 10.0; Win64; x64) "
-            "AppleWebKit/537.36 (KHTML, like Gecko) "
-            "Chrome/91.0.4472.124 Safari/537.36"
-        ),
-        description="User agent string to use for requests.",
-    )
-    timeout: int = Field(
-        default=30,
-        description="Timeout in seconds for HTTP requests.",
-    )
-    max_content_length: int = Field(
-        default=100_000_000,
-        description="Maximum content length in bytes to process.",
-    )
-
-    debug: bool = Field(
-        default=True,
-        description="Boolean flag for debug mode",
-    )
-
-
-class WebScraperToolBase(BaseTool):
-    input_schema = WebpageScraperToolInputSchema
-    output_schema = WebpageScraperToolOutputSchema
-    config_schema = WebpageScraperToolConfig
-
-    async def _extract_author(self, soup: BeautifulSoup) -> Optional[str]:
-        """
-        Extracts the author information from the webpage using multiple methods.
-
-        Args:
-            soup (BeautifulSoup): The parsed HTML content.
-
-        Returns:
-            Optional[str]: The extracted author name(s) or None if not found.
-        """
-        # Check standard meta tags
-        author_meta = soup.find("meta", attrs={"name": "author"})
-        if author_meta and author_meta.get("content"):
-            return author_meta.get("content")
-
-        # Check schema.org markup
-        article_schema = soup.find("script", type="application/ld+json")
-        if article_schema:
-            try:
-                data = json.loads(article_schema.string)
-                if isinstance(data, dict):
-                    author = data.get("author", {}).get("name")
-                    if author:
-                        return author
-            except json.JSONDecodeError:
-                pass
-
-        # Check common author classes/IDs
-        author_elements = soup.find_all(class_=re.compile(r"author|byline", re.I))
-        if author_elements:
-            authors = []
-            for element in author_elements:
-                if element.string and len(element.string.strip()) > 0:
-                    authors.append(element.string.strip())
-            if authors:
-                return ", ".join(authors)
-
-        return None
-
-    async def _extract_metadata(
-        self,
-        soup: BeautifulSoup,
-        doc: Document,
-        url: str,
-    ) -> WebpageMetadata:
-        """
-        Extracts comprehensive metadata from the webpage using multiple methods and sources.
-
-        Args:
-            soup (BeautifulSoup): The parsed HTML content.
-            doc (Document): The readability document.
-            url (str): The URL of the webpage.
-
-        Returns:
-            WebpageMetadata: The extracted metadata.
-        """
-        # domain = urlparse(url).netloc
-
-        metadata = {
-            "query": url,
-            "url": url,
-            "title": doc.title() or self._extract_title(soup),
-            "doi": None,
-            "keywords": [],
-            "author": None,
-            "publication_date": None,
-        }
-
-        # Extract Dublin Core metadata
-        dc_mappings = {
-            "DC.subject": "category",
-            "DC.date.modified": "published_date",
-            "DC.identifier": "doi",  # Some sites use this for DOI
-        }
-
-        for dc_name, meta_key in dc_mappings.items():
-            meta_tag = soup.find("meta", attrs={"name": dc_name})
-            if meta_tag and meta_tag.get("content"):
-                metadata[meta_key] = meta_tag.get("content")
-
-        # Extract Open Graph metadata
-        og_mappings = {
-            "og:description": "description",
-            "og:site_name": "site_name",
-            "og:type": "type",
-            "og:image": "image_url",
-            "og:locale": "language",
-            "og:section": "section",
-            "article:section": "section",
-            "article:tag": "tags",
-            "article:published_time": "published_date",
-            "article:modified_time": "modified_date",
-        }
-
-        for og_prop, meta_key in og_mappings.items():
-            try:
-                meta_tag = soup.find("meta", property=og_prop)
-                if meta_tag and meta_tag.get("content"):
-                    metadata[meta_key] = meta_tag.get("content")
-            except KeyError:
-                continue
-
-        # Extract schema.org metadata (JSON-LD)
-        schema_tags = soup.find_all("script", type="application/ld+json")
-        for tag in schema_tags:
-            try:
-                schema_data = json.loads(tag.string)
-                if isinstance(schema_data, dict):
-                    if "datePublished" in schema_data:
-                        metadata["published_date"] = (
-                            metadata["published_date"] or schema_data["datePublished"]
-                        )
-                    if "keywords" in schema_data:
-                        metadata["keywords"] = (
-                            metadata["keywords"] or schema_data["keywords"]
-                        )
-            except (json.JSONDecodeError, AttributeError):
-                pass
-
-        # Extract Twitter Card metadata
-        twitter_mappings = {
-            "twitter:description": "description",
-            "twitter:site": "site_name",
-        }
-
-        for twitter_name, meta_key in twitter_mappings.items():
-            meta_tag = soup.find("meta", attrs={"name": twitter_name})
-            if meta_tag and meta_tag.get("content"):
-                metadata[meta_key] = metadata[meta_key] or meta_tag.get("content")
-
-        # Extract citation metadata
-        citation_mappings = {
-            "citation_title": "title",
-            "citation_doi": "doi",
-            "citation_journal_title": "publisher",
-            "citation_keywords": "keywords",
-            "citation_publication_date": "published_date",
-        }
-
-        for citation_name, meta_key in citation_mappings.items():
-            meta_tag = soup.find("meta", attrs={"name": citation_name})
-            if meta_key not in metadata:
-                continue
-            if meta_tag and meta_tag.get("content"):
-                metadata[meta_key] = metadata[meta_key] or meta_tag.get("content")
-
-        # Build citation string for academic content
-        if metadata["doi"] or (
-            metadata["title"] and metadata["author"] and metadata["publication_date"]
-        ):
-            citation_parts = []
-            if metadata["author"]:
-                citation_parts.append(metadata["author"])
-            if metadata["title"]:
-                citation_parts.append(f'"{metadata["title"]}"')
-            if "publisher" in metadata:
-                citation_parts.append(metadata["publisher"])
-            if metadata["publication_date"]:
-                citation_parts.append(metadata["publication_date"][:4])  # Just the year
-            if metadata["doi"]:
-                citation_parts.append(f"DOI: {metadata['doi']}")
-
-            if citation_parts:
-                metadata["citation"] = ". ".join(citation_parts)
-
-        # Clean up metadata values
-        for key, value in metadata.items():
-            if isinstance(value, str):
-                metadata[key] = value.strip()
-
-        if isinstance(metadata["keywords"], str):
-            metadata["keywords"] = [metadata["keywords"]]
-        return WebpageMetadata(**metadata)
-
-    async def _extract_title(self, soup: BeautifulSoup) -> str:
-        """
-        Extracts the title using multiple fallback methods.
-
-        Args:
-            soup (BeautifulSoup): The parsed HTML content.
-
-        Returns:
-            str: The extracted title or a default string if not found.
-        """
-        # Try OG title first
-        og_title = soup.find("meta", property="og:title")
-        if og_title:
-            return og_title.get("content")
-
-        # Try regular title tag
-        title_tag = soup.find("title")
-        if title_tag and title_tag.string:
-            return title_tag.string.strip()
-
-        # Try main heading
-        h1_tag = soup.find("h1")
-        if h1_tag and h1_tag.string:
-            return h1_tag.string.strip()
-
-        return None
-
-
-class SimpleWebScraper(WebScraperToolBase):
+class SimpleWebScraper(WebScraper):
     """
     Tool for scraping webpage content and converting it to markdown format.
     """
@@ -314,27 +30,26 @@ class SimpleWebScraper(WebScraperToolBase):
             ValueError: If content length exceeds maximum
             RequestException: For other request-related errors
         """
-        headers = {
-            "User-Agent": self.config.user_agent,
-            "Accept": "text/html,application/xhtml+xml,application/xml;q=0.9,image/webp,*/*;q=0.8",
-            "Accept-Language": "en-US,en;q=0.5",
-            "Connection": "keep-alive",
-        }
-
         try:
-            response = requests.get(url, headers=headers, timeout=self.config.timeout)
-            response.raise_for_status()
+            async with httpx.AsyncClient(
+                timeout=self.timeout,
+                follow_redirects=True,
+            ) as client:
+                response = await client.get(url, headers=self.headers)
+                response.raise_for_status()
 
-            if len(response.content) > self.config.max_content_length:
-                raise ValueError(
-                    f"Content length exceeds maximum of {self.config.max_content_length} bytes",
-                )
+                if len(response.content) > self.max_content_length:
+                    raise ValueError(
+                        f"Content length exceeds maximum of {self.max_content_length} bytes",
+                    )
 
-            return response.text
+                return response.text
 
-        except requests.exceptions.HTTPError as http_err:
+        except httpx.HTTPStatusError as http_err:
             raise HTTPError(f"HTTP error occurred: {http_err}")
-        except requests.exceptions.RequestException as req_err:
+        except httpx.TimeoutException as timeout_err:
+            raise RequestException(f"Request timeout: {timeout_err}")
+        except httpx.RequestError as req_err:
             raise RequestException(f"Error fetching webpage: {req_err}")
 
     async def _clean_markdown(self, markdown: str) -> str:
@@ -389,8 +104,9 @@ class SimpleWebScraper(WebScraperToolBase):
 
     async def _arun(
         self,
-        params: WebpageScraperToolInputSchema,
-    ) -> WebpageScraperToolOutputSchema:
+        params: ScraperToolInputSchema,
+        **kwargs,
+    ) -> ScraperToolOutputSchema:
         """
         Runs the WebpageScraperTool with the given parameters.
 
@@ -403,6 +119,7 @@ class SimpleWebScraper(WebScraperToolBase):
 
         if params.url.path.endswith((".pdf", ".PDF")):
             raise RuntimeError(f"Can't parse url with PDF :: {params.url}")
+
         html_content = await self._fetch_webpage(str(params.url))
 
         # Parse HTML with BeautifulSoup
@@ -434,22 +151,22 @@ class SimpleWebScraper(WebScraperToolBase):
             str(params.url),
         )
 
-        return WebpageScraperToolOutputSchema(
+        return ScraperToolOutputSchema(
             content=markdown_content.strip(),
             metadata=metadata,
         )
 
 
-class Crawl4AIWebScraper(WebScraperToolBase):
+class Crawl4AIWebScraper(WebScraper):
     async def fetch(self, url: str):
         async with AsyncWebCrawler() as crawler:
             return await crawler.arun(url=url)
 
     async def _arun(
         self,
-        params: WebpageScraperToolInputSchema,
+        params: ScraperToolInputSchema,
         **kwargs,
-    ) -> WebpageScraperToolOutputSchema:
+    ) -> ScraperToolOutputSchema:
         """
         Async version of the run method.
         """
@@ -467,184 +184,7 @@ class Crawl4AIWebScraper(WebScraperToolBase):
             str(params.url),
         )
 
-        return WebpageScraperToolOutputSchema(
+        return ScraperToolOutputSchema(
             content=crawl_result.markdown.strip(),
             metadata=metadata,
         )
-
-
-class DoclingScraperConfig(BaseToolConfig):
-    pdf_mode: Literal["accurate", "fast"] = Field(default="fast")
-    export_type: Literal["markdown", "html"] = Field(default="markdown")
-    analyze_image: bool = Field(
-        default=False,
-        description="Use VML to analyze image? (might slow down the parse)",
-    )
-    do_table_structure: bool = Field(default=False)
-    use_ocr: bool = Field(default=False, description="Use OCR for PDF parsing?")
-    allowed_formats: list[InputFormat] = Field(
-        default_factory=lambda: [
-            InputFormat.PDF,
-            InputFormat.DOCX,
-            InputFormat.PPTX,
-            InputFormat.HTML,
-            InputFormat.XLSX,
-            InputFormat.ASCIIDOC,
-            InputFormat.MD,
-        ],
-    )
-    title_fallback: str = Field(
-        default="Untitled",
-    )
-    title_early_search_limit: int = Field(
-        default=10,
-    )
-
-    @field_validator("pdf_mode", "export_type", mode="before")
-    @classmethod
-    def convert_to_lowercase(cls, v):
-        if isinstance(v, str):
-            return v.lower()
-        return v
-
-    @model_validator(mode="after")
-    def add_image_format_if_needed(self):
-        if self.analyze_image and InputFormat.IMAGE not in self.allowed_formats:
-            self.allowed_formats.append(InputFormat.IMAGE)
-        return self
-
-
-class DoclingScraper(WebScraperToolBase):
-    """
-    Scrapes PDF, HTML, docs (and other supported formats) into Markdown using Docling.
-
-    Note:
-
-    - This tool requires the `docling` package to be installed.
-    - It supports both PDF and HTML formats, with options for OCR and table structure parsing.
-    - The `pdf_mode` can be set to "accurate" or "fast"
-    - Make sure `huggingface-cli login` is run before using this tool to access the required models.
-    """
-
-    config_schema = DoclingScraperConfig
-
-    def _post_init(
-        self,
-    ) -> None:
-        super()._post_init()
-        self._setup_converter()
-        self._metadata_extractor = _DoclingMetadataExtractor(
-            early_search_limit=self.config.title_early_search_limit,
-            fallback_title=self.config.title_fallback,
-            debug=self.config.debug,
-        )
-
-    def _setup_converter(
-        self,
-        custom_options: Optional[Dict[str, Any]] = None,
-    ) -> None:
-        """
-        Initialize the DocumentConverter, conditionally configuring for pdf_mode,
-        Common Table Structure (CV) and image analysis options.
-        """
-        # Disable table-structure (CV) if images are not analyzed
-        pipeline_options = PdfPipelineOptions(
-            do_table_structure=self.do_table_structure,
-            do_ocr=self.use_ocr,
-        )
-        pipeline_options.table_structure_options.mode = (
-            TableFormerMode.ACCURATE
-            if self.pdf_mode.lower() == "accurate"
-            else TableFormerMode.FAST
-        )
-
-        format_options: Dict[InputFormat, Any] = {
-            InputFormat.HTML: HTMLFormatOption(),
-            InputFormat.PDF: PdfFormatOption(pipeline_options=pipeline_options),
-        }
-        self.format_options = format_options
-        if custom_options:
-            format_options.update(custom_options)
-
-        if self.debug:
-            logger.debug(
-                f"Docling format options :: {format_options}",
-            )
-
-        self.doc_converter = DocumentConverter(
-            allowed_formats=self.allowed_formats,
-            format_options=format_options,
-        )
-
-    async def _get_docling_document(
-        self,
-        file_path: str,
-        force_format: Optional[InputFormat] = None,
-        reconfigure: bool = False,
-    ) -> DoclingDocument:
-        """
-        Convert a single URL into a DoclingDocument.
-        Raises:
-            RuntimeError      -- conversion failure
-        """
-        if reconfigure or not hasattr(self, "doc_converter"):
-            self._setup_converter()
-
-        try:
-            result = self.doc_converter.convert(file_path)
-            return result.document
-        except Exception as e:
-            raise RuntimeError(f"Conversion failed ({file_path}): {e}")
-
-    async def _clean_markdown(self, md: str) -> str:
-        """
-        Tidy up markdown: collapse excessive blank lines & trim trailing spaces.
-        """
-        md = re.sub(r"\n\s*\n\s*\n", "\n\n", md)
-        lines = [line.rstrip() for line in md.splitlines()]
-        return "\n".join(lines).strip() + "\n"
-
-    async def _process_document(
-        self,
-        path: str,
-    ) -> Tuple[str, WebpageMetadata]:
-        """
-        Full round-trip: fetch doc, export to markdown, clean, and build metadata (docling doesn't support metadata extraction).
-        """
-        doc = await self._get_docling_document(path)
-        markdown = doc.export_to_markdown()
-        markdown = await self._clean_markdown(markdown)
-
-        _docling_metadata = await self._metadata_extractor.arun(doc)
-
-        metadata = WebpageMetadata(
-            url=path,
-            query=path,
-            title=_docling_metadata.title,
-        )
-        return markdown, metadata
-
-    async def _arun(
-        self,
-        params: WebpageScraperToolInputSchema,
-        **kwargs,
-    ) -> WebpageScraperToolOutputSchema:
-        """
-        Entry point for external calls.
-        Returns cleaned markdown + metadata, or raises a descriptive error.
-        """
-        path = str(params.url)
-        try:
-            md, meta = await self._process_document(path)
-            return WebpageScraperToolOutputSchema(content=md, metadata=meta)
-
-        except FileNotFoundError as e:
-            # local file was missing
-            raise RuntimeError(f"[File Not Found] {e}")
-
-        except RuntimeError as e:
-            # issues during conversion
-            raise RuntimeError(f"[Conversion Error] {e}")
-
-        except Exception as e:
-            raise RuntimeError(f"[Internal Error] Failed to scrape {path}") from e
